+++ Grasshopper_Engine/Grasshopper_Engine.csproj
@@ -128,30 +128,12 @@
     <Compile Include="Compute\RenderRhinoWires.cs" />
     <Compile Include="Compute\RenderWires.cs" />
     <Compile Include="Compute\CastToGoo.cs" />
-    <Compile Include="Convert\ToGoo.cs" />
-    <Compile Include="Convert\FromGoo.cs" />
     <Compile Include="Convert\ToNodeParam.cs" />
     <Compile Include="Convert\ToNode.cs" />
     <Compile Include="Convert\ToNodeGroup.cs" />
     <Compile Include="Convert\ToType.cs" />
     <Compile Include="Convert\ToRhino.cs" />
     <Compile Include="Create\DataTree.cs" />
-    <Compile Include="Objects\Types\GH_BHoMAdapter.cs" />
-    <Compile Include="Objects\Types\GH_BakeableObject.cs" />
-    <Compile Include="Objects\Types\GH_Variable.cs" />
-    <Compile Include="Query\AvailableHints.cs" />
-    <Compile Include="Objects\Hints\IGeometryHint.cs" />
-    <Compile Include="Objects\Hints\BHomObjectHint.cs" />
-    <Compile Include="Objects\Hints\DictionaryHint.cs" />
-    <Compile Include="Objects\Hints\EnumHint.cs" />
-    <Compile Include="Objects\Hints\TypeHint.cs" />
-    <Compile Include="Objects\Types\GH_BHoMObject.cs" />
-    <Compile Include="Objects\Types\GH_Dictionary.cs" />
-    <Compile Include="Objects\Types\GH_Enum.cs" />
-    <Compile Include="Objects\Types\GH_IBHoMGeometry.cs" />
-    <Compile Include="Objects\Types\GH_IObject.cs" />
-    <Compile Include="Objects\Types\GH_BHoMGoo.cs" />
-    <Compile Include="Objects\Types\GH_Type.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Query\RenderColour.cs" />
     <Compile Include="Query\RenderMaterial.cs" />
