+++ Grasshopper_UI/Goos/GH_BakeableObject.cs
@@ -34,7 +34,7 @@ using GH_IO;
 using GH_IO.Serialization;
 using BH.Engine.Serialiser;
 
-namespace BH.Engine.Grasshopper.Objects
+namespace BH.UI.Grasshopper.Goos
 {
     public class GH_BakeableObject<T> : GH_BHoMGoo<T>, IGH_PreviewData, IGH_BakeAwareData, GH_ISerializable
     {
@@ -100,7 +100,7 @@ namespace BH.Engine.Grasshopper.Objects
         public override bool CastFrom(object source)
         {
             if (source is IGH_Goo)
-                return CastFrom(Convert.IFromGoo<object>((IGH_Goo)source));
+                return CastFrom(Helpers.IFromGoo<object>((IGH_Goo)source));
 
             if (source.GetType().Namespace.StartsWith("Rhino.Geometry"))
                 source = BH.Engine.Rhinoceros.Convert.FromRhino(source as dynamic);
