+++ Grasshopper_UI/Hints/DictionaryHint.cs
@@ -20,77 +20,34 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-using BH.Engine.Serialiser;
-using GH_IO;
-using GH_IO.Serialization;
-using Grasshopper.Kernel.Types;
 using System;
+using Grasshopper.Kernel.Parameters;
+using BH.UI.Grasshopper.Goos;
 
-namespace BH.Engine.Grasshopper.Objects
+namespace BH.UI.Grasshopper.Hints
 {
-    public class GH_Enum : GH_BHoMGoo<Enum>, GH_ISerializable
+    public class DictionaryHint : IGH_TypeHint
     {
         /*******************************************/
         /**** Properties                        ****/
         /*******************************************/
 
-        public override string TypeName { get; } = "Enum";
-
-        public override string TypeDescription { get; } = "Defines an enum";
-
-        public override bool IsValid { get { return Value != null; } }
+        public Guid HintID { get; } = new Guid("1574563B-80AC-486D-B175-A4F2E5EC76D5"); 
+        public string TypeName { get; } = "Dictionary"; 
 
 
         /*******************************************/
         /**** Constructors                      ****/
         /*******************************************/
 
-        public GH_Enum() : base() { }
-
-        /***************************************************/
-
-        public GH_Enum(Enum val) : base(val) { }
-
-
-        /*******************************************/
-        /**** Override Methods                  ****/
-        /*******************************************/
-
-        public override IGH_Goo Duplicate()
+        public bool Cast(object data, out object target)
         {
-            return new GH_Enum { Value = Value };
-        }
-
-        /*******************************************/
-
-        public override string ToString()
-        {
-            Enum val = Value;
-            if (val == null)
-                return "null";
+            GH_Dictionary dict = new GH_Dictionary() { Value = null };
+            dict.CastFrom(data);
+            if (dict.Value == null)
+                target = data;
             else
-                return val.ToString();
-        }
-
-        /***************************************************/
-
-        public override bool Read(GH_IReader reader)
-        {
-            string json = "";
-            reader.TryGetString("Json", ref json);
-
-            if (json != null && json.Length > 0)
-                Value = (Enum)BH.Engine.Serialiser.Convert.FromJson(json);
-
-            return true;
-        }
-
-        /***************************************************/
-
-        public override bool Write(GH_IWriter writer)
-        {
-            if (Value != null)
-                writer.SetString("Json", Value.ToJson());
+                target = dict.Value;
             return true;
         }
