+++ Grasshopper_UI/Templates/CallerValueList.cs
@@ -27,8 +27,6 @@ using GH = Grasshopper;
 using Grasshopper.Kernel;
 using BH.oM.Base;
 using BH.oM.UI;
-using BH.Engine.Grasshopper;
-using BH.Engine.Grasshopper.Objects;
 using System.Collections.Generic;
 using System.Linq;
 using BH.UI.Base;
