+++ Grasshopper_UI/Templates/DataAccessor_GH.cs
@@ -79,7 +79,7 @@ namespace BH.UI.Grasshopper.Templates
             if (scriptParam != null)
                 hint = scriptParam.SelectedHint;
 
-            return BH.Engine.Grasshopper.Convert.IFromGoo<T>(goo, hint);
+            return Helpers.IFromGoo<T>(goo, hint);
         }
 
         /*************************************/
@@ -96,7 +96,7 @@ namespace BH.UI.Grasshopper.Templates
 
             List<IGH_Goo> goo = new List<IGH_Goo>();
             GH_Accessor.GetDataList<IGH_Goo>(index, goo);
-            return goo.Select(x => BH.Engine.Grasshopper.Convert.FromGoo<T>(x, hint)).ToList();
+            return goo.Select(x => Helpers.FromGoo<T>(x, hint)).ToList();
         }
 
         /*************************************/
@@ -112,7 +112,7 @@ namespace BH.UI.Grasshopper.Templates
                 hint = scriptParam.SelectedHint;
 
             IGH_Param param = Inputs[index];
-            return param.VolatileData.StructureProxy.Select(x => x.Cast<IGH_Goo>().Select(y => BH.Engine.Grasshopper.Convert.IFromGoo<T>(y, hint)).ToList()).ToList();
+            return param.VolatileData.StructureProxy.Select(x => x.Cast<IGH_Goo>().Select(y => Helpers.IFromGoo<T>(y, hint)).ToList()).ToList();
 
             // This shows that using the GetDataTree method from GH is actually giving the exact same result with the exact same problem of collecting the entire data instead of a subtree
             /*IGH_Structure goo = Activator.CreateInstance(typeof(GH_Structure<>).GetGenericTypeDefinition().MakeGenericType(new Type[] { param.Type })) as IGH_Structure;
@@ -170,7 +170,7 @@ namespace BH.UI.Grasshopper.Templates
         private List<List<T>> ConvertDataTree<T, TG>(GH_Structure<TG> structure, int index, List<List<T>> result) where TG : IGH_Goo
         {
             GH_Accessor.GetDataTree(index, out structure);
-            result = structure.Branches.Select(x => x.Select(y => BH.Engine.Grasshopper.Convert.IFromGoo<T>(y)).ToList()).ToList();
+            result = structure.Branches.Select(x => x.Select(y => Helpers.IFromGoo<T>(y)).ToList()).ToList();
             return result;
         }
